% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimmonsMon.R
\name{pokeDex}
\alias{pokeDex}
\title{Function for getting data about a single pokemon}
\usage{
pokeDex(name)
}
\arguments{
\item{name}{name of the pokemon we want information about}
}
\value{
None
}
\description{
Function for getting data about a single pokemon
}
\examples{
pokeDex("Litwick")
}
