% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimmonsMon.R
\docType{data}
\name{pokemon_data}
\alias{pokemon_data}
\title{Data about Pokemon}
\format{
A dataframe wit 1076 number of rows/observations and 47 variables
\describe {
\item{Pokemon.Id}{the ID number for this pokemon in this dataset}
\item{Pokemon.Name}{The official name of the pokemon}
}
}
\source{
2u
}
\usage{
pokemon_data
}
\description{
a dataset with data about pokemon from source
}
\keyword{datasets}
